% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents}
\alias{accidents}
\title{Shunter accidents}
\format{A dataframe with 122 rows and 2 variables:
\describe{
 \item{x}{Number of shunter accidents between 1937 and 1942}
 \item{y}{Number of shunter accidents between 1943 and 1947}
}}
\source{
(Arbous and Kerrich, 1951)
}
\usage{
accidents
}
\description{
The number of accidents incurred by 122 shunters in two consecutive year periods, namely 1937 - 1942 and 1943 - 1947
description
}
\keyword{datasets}
