% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeConstantBCMPests.R
\name{ComputeConstantBCMPests}
\alias{ComputeConstantBCMPests}
\title{Bivariate COM-Poisson Parameter Estimation.}
\usage{
ComputeConstantBCMPests(data, max, startvalues = NULL)
}
\arguments{
\item{data}{A two-column dataset of counts.}

\item{max}{[Set tolerance for precision? Maximum iterations for optimization?].}

\item{startvalues}{A vector of starting values for maximum likelihood estimation. The values are read as follows: c(lambda, nu, p00, p10, p01, p11).
The default is c(1,1, 0.25, 0.25, 0.25, 0.25).}
}
\value{
\code{func_name} will return a list of five elements: $par (Parameter Estimates), $negll (Negative Log-Likelihood), $LRTbpd (Hypothesis Test Statistic),
    $pbpd (Hypothesis Test P-Value), and $se (Standard Errors).
}
\description{
\code{func_name} computes the maximum likelihood estimates of a bivariate COM-Poisson distribution for given count data.
}
\examples{
## Standard usage
data(accidents)
ComputeConstantBCMPests(accidents, 100, c(1.3, .08 , .25 , .25 , .25 , .25))

}
